+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -64,8 +64,10 @@ import java.util.logging.FileHandler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
+
 
 import com.vmware.xenon.common.FileUtils.ResourceEntry;
 import com.vmware.xenon.common.NodeSelectorService.SelectAndForwardRequest;