+++ xenon-common/src/main/java/com/vmware/xenon/services/common/AuthorizationCacheUtils.java
@@ -17,12 +17,17 @@ import java.util.EnumSet;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 
+
+
+import java.util.logging.Level;
+
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.Operation.CompletionHandler;
 import com.vmware.xenon.common.Service;
 import com.vmware.xenon.common.Service.Action;
 import com.vmware.xenon.common.ServiceDocument;
 import com.vmware.xenon.common.ServiceDocumentQueryResult;
+import com.vmware.xenon.common.UriUtils;
 import com.vmware.xenon.common.Utils;
 import com.vmware.xenon.services.common.QueryTask.Query;
 import com.vmware.xenon.services.common.QueryTask.Query.Builder;
@@ -40,15 +45,31 @@ public class AuthorizationCacheUtils {
      * @param s service context to invoke the operation
      * @param op Operation to mark completion/failure
      */
-    public static void clearAuthzCacheForUser(Service s, Operation op) {
+    public static void clearAuthzCacheForUser(Service s,  Operation op) {
         String userPath = op.getUri().getPath();
         op.nestCompletion((o, e) -> {
             if (e != null) {
                 op.fail(e);
                 return;
             }
-            s.getHost().clearAuthorizationContext(s, userPath);
-            op.complete();
+            if (s.getHost().getAuthorizationServiceUri() == null) {
+                op.complete();
+                return;
+            }
+            AuthorizationCacheClearRequest request = AuthorizationCacheClearRequest.create(userPath);
+            Operation postClearCacheRequest = Operation.createPost(s.getHost(), ServiceUriPaths.CORE_AUTHZ_VERIFICATION)
+                    .setBody(request)
+                    .setCompletion((clearOp, clearEx) -> {
+                        if (clearEx != null) {
+                            s.getHost().log(Level.SEVERE, Utils.toString(clearEx));
+                            op.fail(clearEx);
+                            return;
+                        }
+                        op.complete();
+                    });
+            postClearCacheRequest.addPragmaDirective(Operation.PRAGMA_DIRECTIVE_CLEAR_AUTH_CACHE);
+            s.setAuthorizationContext(postClearCacheRequest, s.getSystemAuthorizationContext());
+            s.sendRequest(postClearCacheRequest);
         });
     }
 
@@ -62,7 +83,6 @@ public class AuthorizationCacheUtils {
      * @param userGroupState UserGroup service state
      */
     public static void clearAuthzCacheForUserGroup(Service s, Operation op, UserGroupState userGroupState) {
-
         op.nestCompletion((o, e) -> {
             if (e != null) {
                 op.fail(e);
@@ -89,10 +109,24 @@ public class AuthorizationCacheUtils {
                             op.complete();
                             return;
                         }
+                        AtomicInteger completionCount = new AtomicInteger(0);
+                        CompletionHandler handler = (clearOp, clearEx) -> {
+                            if (clearEx != null) {
+                                s.getHost().log(Level.SEVERE, Utils.toString(clearEx));
+                                op.fail(clearEx);
+                                return;
+                            }
+                            if (completionCount.incrementAndGet() == result.documentLinks.size()) {
+                                op.complete();
+                            }
+                        };
                         for (String userLink : result.documentLinks) {
-                            s.getHost().clearAuthorizationContext(s, userLink);
+                            Operation clearUserOp = new Operation();
+                            clearUserOp.setUri(UriUtils.buildUri(s.getHost(), userLink));
+                            clearUserOp.setCompletion(handler);
+                            clearAuthzCacheForUser(s, clearUserOp);
+                            clearUserOp.complete();
                         }
-                        op.complete();
                     }
                 );
             s.setAuthorizationContext(postOp, s.getSystemAuthorizationContext());