+++ xenon-common/src/main/java/com/vmware/xenon/services/common/ResourceGroupService.java
@@ -80,12 +80,11 @@ public class ResourceGroupService extends StatefulService {
     }
 
     @Override
-    public void handleRequest(Operation request, OperationProcessingStage opProcessingStage) {
-        if (AuthorizationCacheUtils.isAuthzCacheClearApplicableOperation(request)) {
-            AuthorizationCacheUtils.clearAuthzCacheForResourceGroup(this, request);
+    public void processCompletionStageUpdateAuthzArtifacts(Operation op) {
+        if (AuthorizationCacheUtils.isAuthzCacheClearApplicableOperation(op)) {
+            AuthorizationCacheUtils.clearAuthzCacheForResourceGroup(this, op);
         }
-
-        super.handleRequest(request, opProcessingStage);
+        op.complete();
     }
 
     @Override