+++ xenon-common/src/main/java/com/vmware/xenon/services/common/RoleService.java
@@ -82,9 +82,9 @@ public class RoleService extends StatefulService {
     }
 
     @Override
-    public void handleRequest(Operation request, OperationProcessingStage opProcessingStage) {
-        checkAndNestCompletionForAuthzCacheClear(this, request);
-        super.handleRequest(request, opProcessingStage);
+    public void processCompletionStageUpdateAuthzArtifacts(Operation op) {
+        checkAndNestCompletionForAuthzCacheClear(this, op);
+        op.complete();
     }
 
     @Override